backend/requirements.txt
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
openai>=1.30.0
requests==2.32.2
moviepy==1.0.3
numpy<3,>=2.0.0
Pillow<12,>=10.4.0
starlette==0.37.2
ujson==2.0.0
fastapi==0.115.0

